refactor: add global config
