+++ config/color-modes.js
@@ -1,0 +1,3 @@
+const colorModes = [`default`, `light`]
+
+module.exports = colorModes
