+++ config/components.js
@@ -1,0 +1,16 @@
+const components = {
+  codeBlock: {
+    isCopy: true,
+    isFocus: true,
+    isLabel: true,
+  },
+  intro: {
+    id: `introduction`,
+  },
+  tableOfContents: {
+    introId: `introduction`,
+    introTitle: `Introduction`,
+  },
+}
+
+module.exports = components
