+++ config/favicons.js
@@ -1,0 +1,6 @@
+const favicons = {
+  light: `/favicon-light.svg`,
+  dark: `/favicon-dark.svg`,
+}
+
+module.exports = favicons
