+++ config/seo.js
@@ -1,0 +1,23 @@
+const seo = {
+  firstName: `Vladimir`,
+  lastName: `Skok`,
+  siteName: `Vladimir Skok`,
+  logo: {
+    url: `logo.jpg`,
+    width: 112,
+    height: 112,
+  },
+  language: `en`,
+  socialMedia: {
+    twitter: `https://twitter.com/iamskok1`,
+    github: `https://github.com/iamskok`,
+  },
+  address: {
+    addressCountry: `US`,
+    addressLocality: `Atlanta`,
+    addressRegion: `GA`,
+  },
+  speakableSelector: [`[data-speakable="true"]`],
+}
+
+module.exports = seo
