+++ gatsby-config.js
@@ -1,4 +1,4 @@
-const siteMetadata = require("./site-metadata")
+const config = require("./config")
 const forceFileLoader = require("./gatsby/config/force-file-loader")
 const robotsTxt = require("./gatsby/config/robots-txt")
 const manifest = require("./gatsby/config/manifest")
@@ -9,25 +9,8 @@ const sourcePosts = require("./gatsby/config/source-posts")
 const sourceImages = require("./gatsby/config/source-images")
 const sitemap = require("./gatsby/config/sitemap")
 
-require("dotenv").config({
-  path: `.env.${process.env.NODE_ENV}`,
-})
-
-const {
-  NODE_ENV,
-  SITE_URL,
-  URL: NETLIFY_SITE_URL = SITE_URL,
-  DEPLOY_PRIME_URL: NETLIFY_DEPLOY_URL = NETLIFY_SITE_URL,
-  CONTEXT: NETLIFY_ENV = NODE_ENV,
-} = process.env
-const isNetlifyProduction = NETLIFY_ENV === `production`
-const siteUrl = isNetlifyProduction ? NETLIFY_SITE_URL : NETLIFY_DEPLOY_URL
-
 module.exports = {
-  siteMetadata: {
-    ...siteMetadata,
-    siteUrl,
-  },
+  siteMetadata: { ...config },
   plugins: [
     sourceImages,
     sourcePosts,
