+++ gatsby/config/sitemap.js
@@ -1,19 +1,6 @@
+const siteUrl = require("../../config/site-url")
+const pages = require("../../config/pages")
 const slashify = require("../../src/utils/slashify")
-const siteMetadata = require("../../site-metadata")
-
-require("dotenv").config({
-  path: `.env.${process.env.NODE_ENV}`,
-})
-
-const {
-  NODE_ENV,
-  SITE_URL,
-  URL: NETLIFY_SITE_URL = SITE_URL,
-  DEPLOY_PRIME_URL: NETLIFY_DEPLOY_URL = NETLIFY_SITE_URL,
-  CONTEXT: NETLIFY_ENV = NODE_ENV,
-} = process.env
-const isNetlifyProduction = NETLIFY_ENV === `production`
-const siteUrl = isNetlifyProduction ? NETLIFY_SITE_URL : NETLIFY_DEPLOY_URL
 
 const sitemap = {
   resolve: `gatsby-plugin-sitemap`,
@@ -33,7 +20,6 @@ const sitemap = {
     }`,
     resolveSiteUrl: () => siteUrl,
     resolvePages: ({ allMdx: { nodes } }) => {
-      const { pages } = siteMetadata
       const {
         blog: { pathName: blogPathName },
       } = pages
