+++ src/utils/set-theme-favicon.js
@@ -1,6 +1,6 @@
-import { favicons } from "../../site-metadata"
 import setFavicon from "./set-favicon"
 import bustCache from "./bust-cache"
+import favicons from "../../config/favicons"
 
 const setThemeFavicon = () => {
   const { light: faviconLight, dark: faviconDark } = favicons
